feat: Product

+++ database/migrations/2024_11_29_101711_create_produks_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('produks', function (Blueprint $table) {
             $table->id();
             $table->string('nama_produk', 255);
-            $table->decimal("harga", 10, 2);
+            $table->integer("harga");
             $table->integer('stok');
             $table->timestamps();
         });
